moved files to appropriate conventional directories, removed unused codes, functions and classes and intent extras, also rename FurtherVerificationActivity to VerificationActivity
