+++ app/src/main/res/layout/activity_main.xml
@@ -74,7 +74,6 @@
                     android:id="@+id/emailEt"
                     android:layout_marginBottom="10dp"
                     android:inputType="textEmailAddress"
-                    android:text="kassim.bolaji@yahoo.com"
                     />
 
                 <EditText
@@ -85,7 +84,6 @@
                     android:layout_marginBottom="10dp"
                     android:hint="@string/amount_to_charge_required"
                     android:id="@+id/amountEt"
-                    android:text="200"
                     />
 
                 <EditText
@@ -124,7 +122,6 @@
                     android:layout_marginBottom="10dp"
                     android:hint="@string/narration"
                     android:id="@+id/narrationTV"
-                    android:text="jdjjddj"
                     />
 
                 <EditText
@@ -134,7 +131,6 @@
                     android:layout_marginBottom="10dp"
                     android:hint="@string/currency_code_e_g_ngn"
                     android:id="@+id/currencyEt"
-                    android:text="NGN"
                     />
 
                 <EditText
@@ -144,7 +140,6 @@
                     android:layout_marginBottom="10dp"
                     android:hint="@string/country_code_e_g_ng"
                     android:id="@+id/countryEt"
-                    android:text="NG"
                     />
 
                 <EditText
@@ -154,7 +149,6 @@
                     android:layout_marginBottom="10dp"
                     android:hint="@string/first_name"
                     android:id="@+id/fNameEt"
-                    android:text="BKKK"
                     />
 
                 <EditText
@@ -164,7 +158,6 @@
                     android:layout_marginBottom="10dp"
                     android:hint="@string/last_name"
                     android:id="@+id/lnameEt"
-                    android:text="ksksk"
                     />
 
                 <Button
