+++ raveandroid/src/main/java/com/flutterwave/raveandroid/Utils.java
@@ -231,8 +231,4 @@ public class Utils {
         return spacified;
 
     }
-
-    public static void toast(Context context,String msg){
-        Toast.makeText(context, msg, Toast.LENGTH_SHORT).show();
-    }
 }
