+++ raveandroid/src/main/java/com/flutterwave/raveandroid/account/AccountFragment.java
@@ -2,7 +2,6 @@ package com.flutterwave.raveandroid.account;
 
 
 import android.annotation.SuppressLint;
-import android.annotation.TargetApi;
 import android.app.DatePickerDialog;
 import android.app.ProgressDialog;
 import android.content.DialogInterface;
@@ -21,29 +20,24 @@ import android.text.util.Linkify;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.webkit.WebResourceRequest;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
 import android.widget.Button;
 import android.widget.DatePicker;
 import android.widget.EditText;
 import android.widget.FrameLayout;
-import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.flutterwave.raveandroid.FutherVerificationActivity;
+import com.flutterwave.raveandroid.VerificationActivity;
 import com.flutterwave.raveandroid.Payload;
 import com.flutterwave.raveandroid.PayloadBuilder;
 import com.flutterwave.raveandroid.R;
 import com.flutterwave.raveandroid.RavePayActivity;
 import com.flutterwave.raveandroid.RavePayInitializer;
 import com.flutterwave.raveandroid.Utils;
-import com.flutterwave.raveandroid.card.CardPresenter;
 import com.flutterwave.raveandroid.data.Bank;
 import com.flutterwave.raveandroid.data.Callbacks;
-import com.flutterwave.raveandroid.otp_pin_avsvbv_webview.OTPFragment;
-import com.flutterwave.raveandroid.otp_pin_avsvbv_webview.WebFragment;
+import com.flutterwave.raveandroid.OTPFragment;
+import com.flutterwave.raveandroid.WebFragment;
 import com.flutterwave.raveandroid.responses.RequeryResponse;
 
 import java.util.Calendar;
@@ -58,8 +52,6 @@ import static android.view.View.GONE;
  * A simple {@link Fragment} subclass.
  */
 public class AccountFragment extends Fragment implements AccountContract.View, DatePickerDialog.OnDateSetListener {
-
-    public static final String INTENT_SENDER = "accountFrag";
     public static final int FOR_INTERNET_BANKING = 111;
     public static final int FOR_0TP = 222;
     TextInputEditText accountNumberEt;
@@ -68,7 +60,6 @@ public class AccountFragment extends Fragment implements AccountContract.View, D
     Button payButton;
     AccountPresenter presenter;
     Bank selectedBank;
-    BottomSheetBehavior bottomSheetBehaviorOTP;
     TextView pcidss_tv;
     private ProgressDialog progessDialog;
     private BottomSheetDialog bottomSheetDialog;
@@ -79,12 +70,6 @@ public class AccountFragment extends Fragment implements AccountContract.View, D
     private TextInputEditText phoneEt;
     private TextInputLayout phoneTil;
     private String flwRef;
-    private WebView webView;
-    private BottomSheetBehavior bottomSheetBehaviorInternetBanking;
-    private Button otpButton;
-    private TextInputEditText otpEt;
-    private TextInputLayout otpTil;
-    private LinearLayout otpLayout;
     private RavePayInitializer ravePayInitializer;
     private EditText dateOfBirthEt;
     Calendar calendar = Calendar.getInstance();
@@ -105,9 +90,6 @@ public class AccountFragment extends Fragment implements AccountContract.View, D
         // Inflate the layout for this fragment
         View v = inflater.inflate(R.layout.fragment_account, container, false);
 
-        otpTil = (TextInputLayout) v.findViewById(R.id.rave_otpTil);
-        otpEt = (TextInputEditText) v.findViewById(R.id.rave_otpEv);
-        otpButton = (Button) v.findViewById(R.id.rave_otpButton);
         bankEt = (EditText) v.findViewById(R.id.rave_bankEditText);
         amountEt = (TextInputEditText) v.findViewById(R.id.rave_amountTV);
         amountTil = (TextInputLayout) v.findViewById(R.id.rave_amountTil);
@@ -118,7 +100,6 @@ public class AccountFragment extends Fragment implements AccountContract.View, D
         accountNumberEt = (TextInputEditText) v.findViewById(R.id.rave_accountNumberEt);
         accountNumberTil = (TextInputLayout) v.findViewById(R.id.rave_accountNumberTil);
         payButton = (Button) v.findViewById(R.id.rave_payButton);
-        webView = (WebView) v.findViewById(R.id.rave_webview);
         pcidss_tv = (TextView) v.findViewById(R.id.rave_pcidss_compliant_tv);
         dateOfBirthEt = (EditText) v.findViewById(R.id.rave_dobEditText);
 
@@ -132,10 +113,6 @@ public class AccountFragment extends Fragment implements AccountContract.View, D
         Linkify.addLinks(pcidss_tv, pattern, "https://www.pcisecuritystandards.org/pci_security/", null, filter);
 
         FrameLayout internetBankingLayout = (FrameLayout) v.findViewById(R.id.rave_internetBankingBottomSheet);
-        bottomSheetBehaviorInternetBanking = BottomSheetBehavior.from(internetBankingLayout);
-
-        otpLayout = (LinearLayout) v.findViewById(R.id.rave_OTPBottomSheet);
-        bottomSheetBehaviorOTP = BottomSheetBehavior.from(otpLayout);
 
         ravePayInitializer = ((RavePayActivity) getActivity()).getRavePayInitializer();
 
@@ -161,21 +138,6 @@ public class AccountFragment extends Fragment implements AccountContract.View, D
             }
         });
 
-        otpButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String otp = otpEt.getText().toString();
-
-                otpTil.setError(null);
-                otpTil.setErrorEnabled(false);
-
-                if (otp.length() < 1) {
-                    otpTil.setError("Enter a valid one time password");
-                } else {
-                    presenter.validateAccountCharge(flwRef, otp, ravePayInitializer.getPublicKey());
-                }
-            }
-        });
 
         if (Utils.isEmailValid(ravePayInitializer.getEmail())) {
             emailTil.setVisibility(GONE);
@@ -390,17 +352,13 @@ public class AccountFragment extends Fragment implements AccountContract.View, D
     @Override
     public void validateAccountCharge(String pbfPubKey, String flwRef) {
         this.flwRef = flwRef;
-
-        Intent intent = new Intent(getContext(),FutherVerificationActivity.class);
-        intent.putExtra(FutherVerificationActivity.ACTIVITY_MOTIVE,"otp");
-        intent.putExtra(FutherVerificationActivity.INTENT_SENDER,INTENT_SENDER);
+        Intent intent = new Intent(getContext(),VerificationActivity.class);
         startActivityForResult(intent, FOR_0TP);
     }
 
     @Override
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
         if(resultCode == RavePayActivity.RESULT_SUCCESS){
-            //just to be sure this fragment sent the receiving intent
             if(requestCode==FOR_0TP){
                 String otp = data.getStringExtra(OTPFragment.EXTRA_OTP);
                 presenter.validateAccountCharge(flwRef, otp, ravePayInitializer.getPublicKey());
@@ -412,28 +370,10 @@ public class AccountFragment extends Fragment implements AccountContract.View, D
         }
     }
 
-    public boolean closeBottomSheetsIfOpen() {
-        boolean showing = false;
-        if (bottomSheetDialog != null) {
-            if (bottomSheetBehaviorOTP.getState() == BottomSheetBehavior.STATE_EXPANDED) {
-                showing = true;
-                bottomSheetBehaviorOTP.setState(BottomSheetBehavior.STATE_COLLAPSED);
-            }
-
-            if (bottomSheetBehaviorInternetBanking.getState() == BottomSheetBehavior.STATE_EXPANDED) {
-                showing = true;
-                bottomSheetBehaviorInternetBanking.setState(BottomSheetBehavior.STATE_COLLAPSED);
-            }
-        }
-        return showing;
-    }
-
     @Override
     public void onDisplayInternetBankingPage(String authurl, String flwRef) {
         this.flwRef = flwRef;
-        Intent intent = new Intent(getContext(),FutherVerificationActivity.class);
-        intent.putExtra(FutherVerificationActivity.ACTIVITY_MOTIVE,"web");
-        intent.putExtra(FutherVerificationActivity.INTENT_SENDER,INTENT_SENDER);
+        Intent intent = new Intent(getContext(),VerificationActivity.class);
         intent.putExtra(WebFragment.EXTRA_AUTH_URL,authurl);
         startActivityForResult(intent,FOR_INTERNET_BANKING);
     }
@@ -445,20 +385,9 @@ public class AccountFragment extends Fragment implements AccountContract.View, D
 //        showToast(responseAsJSONString);
     }
 
-    private void dismissSheets() {
-        if (bottomSheetBehaviorInternetBanking != null) {
-            bottomSheetBehaviorInternetBanking.setState(BottomSheetBehavior.STATE_COLLAPSED);
-        }
-        if (bottomSheetBehaviorOTP != null) {
-            bottomSheetBehaviorOTP.setState(BottomSheetBehavior.STATE_COLLAPSED);
-        }
-    }
 
     @Override
     public void onPaymentSuccessful(String status, String responseAsJSONString) {
-
-        dismissSheets();
-
         Intent intent = new Intent();
         intent.putExtra("response", responseAsJSONString);
 
@@ -470,7 +399,6 @@ public class AccountFragment extends Fragment implements AccountContract.View, D
 
     @Override
     public void onPaymentFailed(String status, String responseAsJSONString) {
-        dismissSheets();
         Intent intent = new Intent();
         intent.putExtra("response", responseAsJSONString);
         if (getActivity() != null) {
@@ -481,19 +409,16 @@ public class AccountFragment extends Fragment implements AccountContract.View, D
 
     @Override
     public void onValidateSuccessful(String flwRef, String responseAsJsonString) {
-        dismissSheets();
         presenter.requeryTx(flwRef, ravePayInitializer.getSecretKey());
     }
 
     @Override
     public void onValidateError(String message, String responseAsJSonString) {
-        dismissSheets();
         showToast(message);
     }
 
     @Override
     public void onPaymentError(String message) {
-        dismissSheets();
         showToast(message);
     }
