--- raveandroid/src/main/java/com/flutterwave/raveandroid/otp_pin_avsvbv_webview/OTPContract.java
@@ -1,7 +1,0 @@
-package com.flutterwave.raveandroid.otp_pin_avsvbv_webview;
-
-public interface OTPContract {
-    interface View{
-        void showProgressIndicator(boolean active);
-    }
-}
