--- raveandroid/src/main/java/com/flutterwave/raveandroid/otp_pin_avsvbv_webview/OTPPresenter.java
@@ -1,4 +1,0 @@
-package com.flutterwave.raveandroid.otp_pin_avsvbv_webview;
-
-public class OTPPresenter {
-}
