+++ raveandroid/src/main/res/layout/fragment_avsvbv.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".otp_pin_avsvbv_webview.AVSVBVFragment">
+    tools:context=".AVSVBVFragment">
 
     <include layout="@layout/avsvbv_layout"/>
