+++ raveandroid/src/main/res/layout/fragment_card.xml
@@ -198,113 +198,6 @@
 
     </ScrollView>
 
-    <LinearLayout android:layout_width="match_parent"
-        android:orientation="vertical"
-        app:behavior_peekHeight="0dp"
-        android:id="@+id/rave_OTPButtomSheet"
-        android:clickable="true"
-        android:background="#fff"
-        app:behavior_hideable="true"
-        app:layout_behavior="android.support.design.widget.BottomSheetBehavior"
-        android:layout_height="match_parent">
-
-        <View
-            android:layout_width="match_parent"
-            android:layout_marginBottom="40dp"
-            android:layout_height="2dp"
-            style="?OTPHeaderStyle"
-            />
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginBottom="20dp"
-            android:contentDescription="@string/otp_icon"
-            android:src="@drawable/otp"
-            />
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textSize="16sp"
-            android:textColor="#333333"
-            android:layout_marginRight="30dp"
-            android:layout_marginBottom="30dp"
-            android:layout_marginLeft="30dp"
-            android:id="@+id/otp_instructions_tv"
-            android:text="@string/enter_your_one_ntime_password_otp"
-            android:gravity="center_vertical|center_horizontal"
-            />
-
-        <android.support.design.widget.TextInputLayout
-            android:layout_width="match_parent"
-            android:id="@+id/rave_otpTil"
-            android:layout_marginBottom="20dp"
-            android:layout_marginRight="30dp"
-            android:layout_marginLeft="30dp"
-            android:layout_height="wrap_content">
-
-            <android.support.design.widget.TextInputEditText
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableLeft="@drawable/ic_lock_outline_black_24dp"
-                android:drawableStart="@drawable/ic_lock_outline_black_24dp"
-                android:hint="@string/otp"
-                android:imeOptions="actionDone"
-                android:inputType="number"
-                android:drawablePadding="10dp"
-                android:id="@+id/rave_otpEv"
-                />
-
-        </android.support.design.widget.TextInputLayout>
-
-        <Button
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/rave_otpButton"
-            style="?OTPButtonStyle"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginRight="30dp"
-            android:layout_marginLeft="30dp"
-            android:text="@string/enter"
-            />
-
-    </LinearLayout>
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:behavior_peekHeight="0dp"
-        android:id="@+id/rave_VBVBottomSheet"
-        android:background="#fff"
-        android:clickable="true"
-        app:behavior_hideable="true"
-        app:layout_behavior="android.support.design.widget.BottomSheetBehavior"
-        android:orientation="vertical">
-
-        <WebView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:id="@+id/rave_webview"/>
-
-        <FrameLayout
-            android:layout_width="match_parent"
-            android:background="#372E4C"
-            android:visibility="gone"
-            android:id="@+id/rave_progressContainer"
-            android:layout_height="match_parent">
-
-            <ProgressBar
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center_horizontal|center_vertical"/>
-
-
-        </FrameLayout>
-
-
-    </FrameLayout>
 
 </android.support.design.widget.CoordinatorLayout>
